chat_bot_env/requirements.txt
aiogram==3.22.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
apscheduler==3.10.4
